+++ src/main/java/com/li64/tide/datagen/fabric/providers/assets/TideModelProvider.java
@@ -68,8 +68,10 @@ public class TideModelProvider extends FabricModelProvider {
         generator.generateFlatItem(TideItems.FISHING_JOURNAL, ModelTemplates.FLAT_ITEM);
         generator.generateFlatItem(TideItems.FISHY_NOTE, ModelTemplates.FLAT_ITEM);
 
-        generator.generateFlatItem(TideItems.CLIMATE_GAUGE, ModelTemplates.FLAT_ITEM);
         generator.generateFlatItem(TideItems.POCKET_WATCH, ModelTemplates.FLAT_ITEM);
+        generator.generateFlatItem(TideItems.LUNAR_CALENDAR, ModelTemplates.FLAT_ITEM);
+        generator.generateFlatItem(TideItems.CLIMATE_GAUGE, ModelTemplates.FLAT_ITEM);
+        generator.generateFlatItem(TideItems.DEPTH_METER, ModelTemplates.FLAT_ITEM);
 
         generator.generateFlatItem(TideItems.ENCHANTED_POCKET_WATCH, ModelTemplates.FLAT_ITEM);
         generator.generateFlatItem(TideItems.DRAGONFIN_BOOTS, ModelTemplates.FLAT_ITEM);