+++ src/main/java/com/li64/tide/mixin/AbstractFishMixin.java
@@ -3,12 +3,10 @@ package com.li64.tide.mixin;
 import com.li64.tide.data.FishLengthHolder;
 import com.li64.tide.data.fishing.FishData;
 import net.minecraft.nbt.CompoundTag;
-import net.minecraft.util.RandomSource;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.animal.AbstractFish;
 import net.minecraft.world.item.ItemStack;
 import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
@@ -57,7 +55,7 @@ public abstract class AbstractFishMixin implements FishLengthHolder {
     private void tide$randomizeLength() {
         Entity self = (Entity)(Object)this;
         this.tide$length = FishData.get(self)
-                /*? if >=1.21.1 {*/.map(data -> data.getRandomLength(self.getRandom())
+                /*? if >=1.21.1 {*/.map(data -> data.getRandomLength(self.getRandom()))
                 /*?} else*//*.map(data -> data.getRandomLength(self.random))*/
                 .orElse(0.0);
     }