+++ src/main/java/com/li64/tide/registries/TideItems.java
@@ -90,8 +90,10 @@ public class TideItems {
             /*? if >=1.21*/.component(DataComponents.BUNDLE_CONTENTS, new BundleContents(List.of()))
             .stacksTo(1));
 
-    public static final Item CLIMATE_GAUGE = register("climate_gauge", ClimateGaugeItem::new);
     public static final Item POCKET_WATCH = register("pocket_watch", PocketWatchItem::new);
+    public static final Item LUNAR_CALENDAR = register("lunar_calendar", LunarCalendarItem::new);
+    public static final Item CLIMATE_GAUGE = register("climate_gauge", ClimateGaugeItem::new);
+    public static final Item DEPTH_METER = register("depth_meter", DepthMeterItem::new);
 
     public static final Item ENCHANTED_POCKET_WATCH = register("enchanted_pocket_watch", EnchantedPocketWatchItem::new, new Item.Properties()
             /*? if >=1.21*/.component(DataComponents.ENCHANTMENT_GLINT_OVERRIDE, true)
@@ -154,7 +156,7 @@ public class TideItems {
 
                     BAIT, LUCKY_BAIT, MAGNETIC_BAIT,
 
-                    CLIMATE_GAUGE, POCKET_WATCH,
+                    POCKET_WATCH, LUNAR_CALENDAR, CLIMATE_GAUGE, DEPTH_METER,
 
                     ENCHANTED_POCKET_WATCH, STARLIGHT_BOW, MIDAS_FISHING_ROD, DRAGONFIN_BOOTS,