+++ src/main/java/com/li64/tide/registries/items/PocketWatchItem.java
@@ -2,43 +2,17 @@ package com.li64.tide.registries.items;
 
 import com.li64.tide.Tide;
 import com.li64.tide.util.TideUtils;
-import net.minecraft.ChatFormatting;
 import net.minecraft.network.chat.Component;
-import net.minecraft.network.chat.Style;
-import net.minecraft.network.protocol.game.ClientboundSetActionBarTextPacket;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.server.level.ServerPlayer;
-import net.minecraft.sounds.SoundEvents;
-import net.minecraft.sounds.SoundSource;
-import net.minecraft.world.InteractionHand;
-import net.minecraft.world.InteractionResultHolder;
-import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.level.Level;
-import org.jetbrains.annotations.NotNull;
 
-import java.util.function.Consumer;
-
-public class PocketWatchItem extends SimpleTooltipItem {
+public class PocketWatchItem extends AbstractProfilingItem {
     public PocketWatchItem(Properties properties) {
-        super(properties);
+        super(properties, Component.translatable("item.tide.pocket_watch.desc"));
     }
 
     @Override
-    public @NotNull InteractionResultHolder<ItemStack> use(@NotNull Level level, @NotNull Player player, @NotNull InteractionHand hand) {
-        if (player instanceof ServerPlayer serverPlayer) {
-            long time = level.getDayTime();
-            serverPlayer.connection.send(new ClientboundSetActionBarTextPacket(
-                    Component.literal(TideUtils.ticksToRealTime(time, Tide.CONFIG.journal.useAmPm))));
-            serverPlayer.level().playSound(null, serverPlayer.blockPosition(),
-                    SoundEvents.STONE_BUTTON_CLICK_ON, SoundSource.PLAYERS, 1.0f, 2.0f);
-        }
-        return InteractionResultHolder.sidedSuccess(player.getItemInHand(hand), level.isClientSide());
-    }
-
-    @Override
-    public void addTooltip(ItemStack stack, Consumer<Component> tooltip) {
-        Style gray = Component.empty().getStyle().withColor(ChatFormatting.GRAY);
-        tooltip.accept(Component.translatable("item.tide.pocket_watch.desc").setStyle(gray));
+    public Component getDisplayedInfo(ServerLevel level, ServerPlayer player) {
+        return Component.literal(TideUtils.ticksToRealTime(level.getDayTime(), Tide.CONFIG.journal.useAmPm));
     }
 }